make everything compile with visual studio 7.1, and in debug mode in win32.
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code@82 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/CAnimatedMeshMS3D.cpp
@@ -357,7 +357,7 @@ bool CAnimatedMeshMS3D::loadFile(io::IReadFile* file)
 
 			//look, if we already have this vertex in our vertex array
 			s32 index = -1;
-			for (int iV = 0; iV < Vertices.size(); iV++)
+			for (int iV = 0; iV < (s32)Vertices.size(); iV++)
 			{
 				if (v == Vertices[iV])
 				{
